improved file organization
